+++ astroquery/ipac/irsa/__init__.py
@@ -4,7 +4,8 @@ IRSA Query Tool
 ===============
 
 This module contains various methods for querying the
-IRSA Catalog Query Service(CatQuery).
+IRSA Catalog Query Service(CatQuery) and the Moving
+Object Search Tool (MOST).
 """
 from astropy import config as _config
 
@@ -14,12 +15,19 @@ class Conf(_config.ConfigNamespace):
     Configuration parameters for `astroquery.ipac.irsa`.
     """
 
-    server = _config.ConfigItem(
+    irsa_server = _config.ConfigItem(
         'https://irsa.ipac.caltech.edu/cgi-bin/Gator/nph-query',
         'Name of the IRSA mirror to use.')
     gator_list_catalogs = _config.ConfigItem(
         'https://irsa.ipac.caltech.edu/cgi-bin/Gator/nph-scan',
         'URL from which to list all the public catalogs in IRSA.')
+    most_server = _config.ConfigItem(
+        'https://irsa.ipac.caltech.edu/cgi-bin/MOST/nph-most',
+        'URL address of the MOST service.')
+    most_interface_url = _config.ConfigItem(
+        'https://irsa.ipac.caltech.edu/applications/MOST/',
+        'URL address of the MOST application interface.'
+    )
     row_limit = _config.ConfigItem(
         500,
         'Maximum number of rows to retrieve in result')
@@ -32,5 +40,6 @@ conf = Conf()
 
 
 from .core import Irsa, IrsaClass
+from .most import Most, MostClass
 
-__all__ = ['Irsa', 'IrsaClass', 'Conf', 'conf', ]
+__all__ = ['Irsa', 'IrsaClass', 'Most', 'MostClass', 'Conf', 'conf', ]
